+++ app/Models/RequestLog.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\DB;
 
 /**
  * Request log
@@ -25,5 +26,19 @@ class RequestLog extends Model
     public function paginatedCollection(int $offset = 0, int $limit = 10)
     {
         return $this->orderByDesc('created_at')->offset($offset)->limit($limit)->get();
+    }
+
+    public function monthlyRequests()
+    {
+        $collection = $this->orderBy(DB::raw("DATE_FORMAT(`request_log`.`created_at`, '%Y-%m')"))
+            ->groupBy(DB::raw("DATE_FORMAT(`request_log`.`created_at`, '%Y-%m')"))
+            ->select(
+                DB::raw("DATE_FORMAT(`request_log`.`created_at`, '%Y-%m')"),
+                DB::raw("COUNT(`request_log`.`id`) AS `requests`"),
+                DB::raw("ANY_VALUE(DATE_FORMAT(`request_log`.`created_at`, '%Y')) AS `year`"),
+                DB::raw("ANY_VALUE(DATE_FORMAT(`request_log`.`created_at`, '%M')) AS `month`")
+            );
+
+        return $collection->get();
     }
 }