+++ routes/api/no-middleware.php
@@ -12,6 +12,8 @@ Route::group(
         Route::options('request/error-log', 'RequestController@optionsErrorLog');
         Route::get('request/log', 'RequestController@log');
         Route::options('request/log', 'RequestController@optionsLog');
+        Route::get('request/log/monthly-requests', 'RequestController@monthlyRequests');
+        Route::options('request/log/monthly-requests', 'RequestController@optionsMonthlyRequests');
 
         Route::post('request/error-log', 'RequestController@createErrorLog');
     }